// src/api/models/Policyholder.ts
export interface Policyholder {
    /**
     * 
     * @type {Person}
     * @memberof Policyholder
     */
    personalInformation?: Person;
    /**
     * Vorsteuerabzug ja/nein
     * @type {string}
     * @memberof Policyholder
     */
    inputTaxDeduction?: PolicyholderInputTaxDeductionEnum;
    /**
     * Fahrzeugmarke || will be extracted from <Policy Pass>
     * @type {string}
     * @memberof Policyholder
     */
    vehicleMake?: string;
    /**
     * Fahrzeugmodell || will be extracted from <Policy Pass>
     * @type {string}
     * @memberof Policyholder
     */
    vehicleType?: string;
    /**
     * Kennzeichen ||will be extracted from <Policy Pass>
     * @type {string}
     * @memberof Policyholder
     */
    vehicleReg?: string;
    /**
     * Versicherungsanbieter || will be extracted from <Companion Pass>
     * @type {string}
     * @memberof Policyholder
     */
    insuranceCompany?: string;
    /**
     * Policennummer || will be extracted from <Policy Pass>
     * @type {string}
     * @memberof Policyholder
     */
    policyNumber?: string;
    /**
     * Fahrzeug-Identifizierungsnummer || will be extracted from <Policy Pass>
     * @type {string}
     * @memberof Policyholder
     */
    vin?: string;
    /**
     * aktueller Kilometerstand
     * @type {number}
     * @memberof Policyholder
     */
    currentMileage?: number;
    /**
     * Grüne-Karte Nummer
     * @type {string}
     * @memberof Policyholder
     */
    greencardNumber?: string;
    /**
     * Ablaufdatum der Grünen Karte
     * @type {Date}
     * @memberof Policyholder
     */
    greencardExpirydate?: Date;
    /**
     * Vollkaskoversichert ja/nein
     * @type {string}
     * @memberof Policyholder
     */
    comprehensiveInsurance?: PolicyholderComprehensiveInsuranceEnum;
}


/**
 * @export
 */
export const PolicyholderInputTaxDeductionEnum = {
    NotSpecified: 'not_specified',
    False: 'false',
    True: 'true'
} as const;
export type PolicyholderInputTaxDeductionEnum = typeof PolicyholderInputTaxDeductionEnum[keyof typeof PolicyholderInputTaxDeductionEnum];

/**
 * @export
 */
export const PolicyholderComprehensiveInsuranceEnum = {
    NotSpecified: 'not_specified',
    False: 'false',
    True: 'true'
} as const;
export type PolicyholderComprehensiveInsuranceEnum = typeof PolicyholderComprehensiveInsuranceEnum[keyof typeof PolicyholderComprehensiveInsuranceEnum];